Work in progress: added to TrackRecordingActivity...
