+++ src/main/java/de/dennisguse/opentracks/TrackRecordingActivity.java
@@ -2,7 +2,6 @@ package de.dennisguse.opentracks;
 
 import android.app.AlertDialog;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener;
 import android.os.Build;
 import android.os.Bundle;
@@ -17,6 +16,7 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 import androidx.viewpager2.adapter.FragmentStateAdapter;
 
+import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.tabs.TabLayoutMediator;
 
 import java.util.List;
@@ -31,6 +31,7 @@ import de.dennisguse.opentracks.fragments.ChooseActivityTypeDialogFragment;
 import de.dennisguse.opentracks.fragments.StatisticsRecordingFragment;
 import de.dennisguse.opentracks.services.TrackRecordingService;
 import de.dennisguse.opentracks.services.TrackRecordingServiceConnection;
+import de.dennisguse.opentracks.services.handlers.GpsStatusValue;
 import de.dennisguse.opentracks.settings.PreferencesUtils;
 import de.dennisguse.opentracks.settings.SettingsActivity;
 import de.dennisguse.opentracks.ui.intervals.IntervalsFragment;
@@ -70,48 +71,44 @@ public class TrackRecordingActivity extends AbstractActivity implements ChooseAc
 
     private TrackRecordingService.RecordingStatus recordingStatus = TrackRecordingService.STATUS_DEFAULT;
 
-    private final TrackRecordingServiceConnection.Callback bindChangedCallback = new TrackRecordingServiceConnection.Callback() {
-        @Override
-        public void onConnected(TrackRecordingService service) {
+    private final TrackRecordingServiceConnection.Callback bindChangedCallback = service -> {
+        service.getRecordingStatusObservable()
+                .observe(TrackRecordingActivity.this, this::onRecordingStatusChanged);
 
-            service.getRecordingStatusObservable()
-                    .observe(TrackRecordingActivity.this, status -> onRecordingStatusChanged(status));
+        service.getGpsStatusObservable()
+                .observe(TrackRecordingActivity.this, this::onGpsStatusChanged);
 
-            if (!service.isRecording()) {
-                if (trackId == null) {
-                    // trackId isn't initialized -> leads a new recording.
-                    trackId = service.startNewTrack();
-                } else {
-                    // trackId is initialized -> resumes the track.
-                    service.resumeTrack(trackId);
-                }
-
-                // A recording track is on.
-                trackDataHub.loadTrack(trackId);
-                trackDataHub.setRecordingStatus(recordingStatus);
+        if (!service.isRecording()) {
+            if (trackId == null) {
+                // trackId isn't initialized -> leads a new recording.
+                trackId = service.startNewTrack();
+            } else {
+                // trackId is initialized -> resumes the track.
+                service.resumeTrack(trackId);
             }
+
+            // A recording track is on.
+            trackDataHub.loadTrack(trackId);
+            trackDataHub.setRecordingStatus(recordingStatus);
         }
     };
 
-    private final OnSharedPreferenceChangeListener sharedPreferenceChangeListener = new OnSharedPreferenceChangeListener() {
-        @Override
-        public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-            if (PreferencesUtils.isKey(R.string.stats_show_on_lockscreen_while_recording_key, key)) {
-                setLockscreenPolicy();
-            }
-            if (PreferencesUtils.isKey(R.string.stats_keep_screen_on_while_recording_key, key)) {
-                setScreenOnPolicy();
-            }
-            if (PreferencesUtils.isKey(R.string.stats_fullscreen_while_recording_key, key)) {
-                setFullscreenPolicy();
-            }
-            if (PreferencesUtils.isKey(R.string.recording_distance_interval_key, key)) {
-                trackDataHub.setRecordingDistanceInterval(PreferencesUtils.getRecordingDistanceInterval());
-            }
-            if (key == null) return;
-
-            runOnUiThread(TrackRecordingActivity.this::invalidateOptionsMenu);
+    private final OnSharedPreferenceChangeListener sharedPreferenceChangeListener = (sharedPreferences, key) -> {
+        if (PreferencesUtils.isKey(R.string.stats_show_on_lockscreen_while_recording_key, key)) {
+            setLockscreenPolicy();
         }
+        if (PreferencesUtils.isKey(R.string.stats_keep_screen_on_while_recording_key, key)) {
+            setScreenOnPolicy();
+        }
+        if (PreferencesUtils.isKey(R.string.stats_fullscreen_while_recording_key, key)) {
+            setFullscreenPolicy();
+        }
+        if (PreferencesUtils.isKey(R.string.recording_distance_interval_key, key)) {
+            trackDataHub.setRecordingDistanceInterval(PreferencesUtils.getRecordingDistanceInterval());
+        }
+        if (key == null) return;
+
+        runOnUiThread(TrackRecordingActivity.this::invalidateOptionsMenu);
     };
 
     private MenuItem insertMarkerMenuItem;
@@ -405,5 +402,19 @@ public class TrackRecordingActivity extends AbstractActivity implements ChooseAc
 
         setLockscreenPolicy();
         setScreenOnPolicy();
+    }
+
+    private void onGpsStatusChanged(GpsStatusValue gpsStatusValue) {
+        Log.e("probando", "onGpsStautsChanged: is started? " + gpsStatusValue.isGpsStarted() + " | " + gpsStatusValue);
+        var coordinatorLayout = viewBinding.trackRecordingCoordinatorLayout;
+        if (coordinatorLayout == null) {
+            return;
+        }
+        Snackbar snackbar = Snackbar
+                .make(coordinatorLayout, "GPS Status changed: " + gpsStatusValue, Snackbar.LENGTH_INDEFINITE)
+                .setAction("OK", v -> {
+
+                });
+        snackbar.show();
     }
 }
